What didn't I change....
I modded a whole bunch this time around.
First off, I created a proper new game command that is almost done,
meaning that all of the values asked for are completely pointless.

Secondly, I added some extreme functionality to the import command,
which imports an external scenario file.

I need to add a delete command for xml's or have an option to open the
folder, or have an option to replace it, but that could add another
parser into the mix.