got schemas working
schemas work but barely. the main document works and validates but sub
elements don't. I need to figure out a way to get the sub elements to
validate properly or figure out a stricter way of generating schemas.
But this is something to implement later on. I may just put this in now
in order to get a small baseline check to see if the file is valid, in
order to make sure not just anything is loaded into the game. Besides
that, some things are working out pretty good, this project is 4 or 5
days away from reaching completion, so there's that.