Made save overwriting work
Resolved an issue where the name of the save file was not being modified
when being parsed by saveParser.  The event that proceeded upon trying
to save in specific scenarios was a catastrophe. If one was to try to
overwrite their save, theyd be greeted with the general catch-all
response of "there was an error saving". Luckily after some CSI , the
suspect was identified: a null valued SaveName variable.

I was quick to set him back in his proper ways. I gave him a proper home
and told him his name so that he could refer to himself henceforth
without a fear of being nameless and transparent to the rest of the
code.