Settings now saves
^^ that, I was able to get the settings to save and to work without fail
(so far). Getting tired now but I still need to power through this. I
also got the Terminal_GUI to display "stats" and a "minimap" but those
are just placeholders for the actual things. Might just zone out after
this one for a few mins. I still need to implement a few more commands
and this game should be ready to go.