And I'm Back Baby!!!
So, I decided to pick back up the “Claw Of Kraymoar” project to kind of
remaster and officially release it. The majority of the art will stay
the same as well as the majority of the storyline, but the game’s
functionality, stability and overall UI appeal would be greatly
improved. This is also being written at around 2:23 AM so please excuse
any spelling / grammatical errors. Anyways, this will hopefully be done
within the next couple of weeks, as the tutu date finally arrives. —JPEG