Updated the resolution selector
This is really taking awhile to get right, due to the fact that I have
netflix and youtube on my computer and I need time in between problems
to sort them out. This new UI is frustrating, but I seem to be learning
slightly. Its 2:30 AM again and its time for me to go to bed, hopefully
I will be able to make a breakthrough in the UI sometime tomorrow.