Completed the main menu
While I do have to learn a shit load more for the new GUI system in
unity, I think that I know enough to move on to the actual game now,
well at least some of the GUI elements for it at least. Anyways, I have
decided that this version of the game will have 3 save files, in the
traditional Metroid / Zelda fashion. Also, a small note to self, do a
clean fucking run of the game and exit it before you push a release out
into the wild for mac (at least I think the problem only exists on
mac). more details are in the code for the main menu. Goodnight for
now, its time for me to go to sleep… at 4:10 A.M..