started new game UI
about to pass out so I’ll make this quick. I’m currently working on a
new game UI that is also a load game UI (its tricky but I think that I
can do it I may have to break it up tho). The majority of the code so
far is a jumbled mess. Once I get some sleep, I will be able to
actually do something.