Updated the GUI and more
Fixed some weird bugs dealing with the length of input and the way the
scroll area is resized. I also fixed a weird thing with the enter key,
but thats all fixed now. I should be working on the parser sometime
soon….

I Just realized that I never committed this… doing another one on top,
well bottom.

Ok updated a whole bunch of stuff dealing with the new game parser. I
decided to add in a bit of tutorial stuff to make getting used to the
game easier / make the game transition better… end with open door.