Fixed weird bugs with the code
Yeah, this one was kinda weird. For some reason the return key and the
enter key on mac / windows respectively output something different. The
macs don’t add a newline while windows systems do. So, the fix: change
the acceptable input to name and all was fine. Thanks Eric and TJ for
helping me debug this one via video chat.