Improved Cinematics
I think that I have successfully created a cinematic system for Claw Of
Kraymoar. Well, Its more of a system that breaks down large text
narrations and adds a "continue" button to them. Anyways, the game is
slowly moving along with several components that could be re created for
future use. I still have to get the save and load functions to work as
well as properly populating the newgame and load menus. I'm done for
tonight and I need to sleep. Hopefully I will be able to work on this
more tomorrow.