Added some stuff
added in a cutscene for the game starting. I need to find a way to tell
the system if the game is loading. I also need to figure out how the
command structure is going to work for the commands, but I may default
to something smaller or go back to the old method. Anyways, venture bros
special is on and I need to watch that.