Added commands / subcommands
the commands dont do anything yet (besides help) unless something has
subcommands, in which it displays all subcommands if a command is
entered without anything else (if they do have subcommands). anyways, I
also changed the font on all of the terminals to be Lucida Sans Console.
This allows things to work better, like much better. I also changed "The
Claw of Kraymoar" to have comic sans instead of another font, because I
think that its hilarious. I also added in an icon for the game, which is
a picture of Aaron Kramer, because its hilarious.