Four or so commands completed
I have completed four or so commands for the game. They still need to be
extensively tested but I'm banking on them working tomorrow morning.
Help is working no problem, same thing with clear. quit may have some
issues but I still need to have someone try to break it. Look is almost
complete but it still needs the object search function added (But I need
to create that first).